Use a tighter ObjectId representation to reduce memory usage
On a modern HotSpot JVM an object header costs 8 bytes of memory, plus
any fields of the object.  Arrays cost 12 bytes of memory (8 bytes for
the header, 4 bytes for the length) plus whatever space is required
for the entries of the array.

This means our prior byte[] ObjectId representation was costing us
8 + 4 + 12 + 20 = 44 bytes on a 32 bit JVM and possibly 48 bytes on
a 64 bit JVM.  This is more than twice the size of the actual data
we are trying to represent here.

By converting each SHA-1 from 20 individual bytes in an array and
storing them as 5 integer values we can reduce the memory footprint
for an ObjectId down to 28 bytes per instance (possibly rounded up
to 32 bytes to ensure 8 byte alignment during allocation).

This conversion comes with a runtime performance penalty, as Java is
simply not able to convert an aligned sequence of 4 bytes into a 4
byte integer via typecasting like C can.  Instead we must do bit
shifts and bitwise OR operations, sucking down CPU time pretty hard.

The hex parse and format routines were hit especially hard so I
have switched them to use table based parsers and lookups.  The
table based version is much faster then the algorithmic variant
we had been using up until now.

The comparsion and parsing routines were hand micro-optimized to get
the best throughput we can from the Sun HotSpot JIT.  I did a lot of
testing running the code in hard tight loops to come up with these
particular combinations of Java statements.  Anything else faired a
lot worse, as the JIT was simply not able to generate better code.

With the optimizated compare and parse methods, as well as the smaller
memory footprint, this version of ObjectId is actually benchmarking
faster than the prior byte[] version when used within RevWalk, and it
has a lower memory overhead.  The additional code complexity is worth
the runtime improvement.

As part of this change I cleaned up the API of ObjectId so it is more
consistent regarding how we create instances from hex strings, from
either "byte[] US-ASCII encoding" or a standard String representation.

+++ org.spearce.jgit/src/org/spearce/jgit/lib/PackIndexV1.java
@@ -62,19 +62,19 @@ class PackIndexV1 extends PackIndex {
 		byte[] data = idxdata[levelOne];
 		if (data == null)
 			return -1;
-		long high = data.length / (4 + Constants.OBJECT_ID_LENGTH);
-		long low = 0;
+		int high = data.length / (4 + Constants.OBJECT_ID_LENGTH);
+		int low = 0;
 		do {
-			final long mid = (low + high) / 2;
-			final long pos = ((4 + Constants.OBJECT_ID_LENGTH) * mid) + 4;
+			final int mid = (low + high) / 2;
+			final int pos = ((4 + Constants.OBJECT_ID_LENGTH) * mid) + 4;
 			final int cmp = objId.compareTo(data, pos);
 			if (cmp < 0)
 				high = mid;
 			else if (cmp == 0) {
-				int b0 = data[(int)pos-4] & 0xff;
-				int b1 = data[(int)pos-3] & 0xff;
-				int b2 = data[(int)pos-2] & 0xff;
-				int b3 = data[(int)pos-1] & 0xff;
+				int b0 = data[pos-4] & 0xff;
+				int b1 = data[pos-3] & 0xff;
+				int b2 = data[pos-2] & 0xff;
+				int b3 = data[pos-1] & 0xff;
 				return (((long)b0) << 24) | ( b1 << 16 ) | ( b2 << 8 ) | (b3);
 			} else
 				low = mid + 1;
